third_party/WebKit/LayoutTests/http/tests/security/suborigins/crossorigin/suborigin-cross-origin-script-window-onerror-no-cors-expected.txt

CONSOLE ERROR: line 1: Uncaught SomeError
This is a testharness.js-based test.
PASS Verify that thrown exception is sanitized when crossorigin script loaded without crossorigin attribute 
Harness: the test ran to completion.
